Move AzleTypeRef back into TsAst
