+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/mod.rs
@@ -7,12 +7,9 @@ pub mod func;
 pub mod opt;
 pub mod record;
 pub mod service;
-pub mod ts_ast;
 pub mod tuple;
 pub mod type_alias;
 pub mod variant;
-
-pub use ts_ast::AzleTypeRef;
 
 impl TsAst {
     pub fn build_candid_types(&self) -> CandidTypes {
